LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/opacity-expected.txt


PASS fillOpacity responsive to inherited changes
FAIL floodOpacity responsive to inherited changes assert_equals: expected "0.375" but got "0.75"
FAIL opacity responsive to inherited changes assert_equals: expected "0.375" but got "0.75"
FAIL shapeImageThreshold responsive to inherited changes assert_equals: expected "0.375" but got "0.75"
FAIL stopOpacity responsive to inherited changes assert_equals: expected "0.375" but got "0.75"
PASS strokeOpacity responsive to inherited changes
